move lots of stuff - some steps to pipelines/steps, internal tool containers to internal tools, reports to own module and connector report to pipelines/connectors
