+++ airbyte-ci/connectors/pipelines/pipelines/dagger/actions/secrets.py
@@ -10,11 +10,11 @@ from typing import TYPE_CHECKING, Callable
 
 from dagger import Container, Secret
 from pipelines.helpers.utils import get_file_contents, get_secret_host_variable
-from pipelines.pipeline.connectors.context import PipelineContext
 
 if TYPE_CHECKING:
     from dagger import Container
     from pipelines.pipeline.connectors.context import ConnectorContext
+    from pipelines.pipeline.connectors.context import PipelineContext
 
 
 async def get_secrets_to_mask(ci_credentials_with_downloaded_secrets: Container) -> list[str]:
