+++ airbyte-ci/connectors/pipelines/pipelines/helpers/github.py
@@ -11,7 +11,7 @@ from typing import TYPE_CHECKING, Optional
 
 from connector_ops.utils import console
 from pipelines import main_logger
-from pipelines.models.bases import CIContext
+from pipelines.models.contexts import CIContext
 
 if TYPE_CHECKING:
     from logging import Logger
