+++ airbyte-ci/connectors/pipelines/pipelines/models/bases.py
@@ -6,9 +6,7 @@
 
 from __future__ import annotations
 
-import json
 import logging
-import webbrowser
 from abc import ABC, abstractmethod
 from dataclasses import dataclass, field
 from datetime import datetime, timedelta
@@ -18,20 +16,12 @@ from typing import TYPE_CHECKING, Any, ClassVar, List, Optional, Set
 import anyio
 import asyncer
 from anyio import Path
-from connector_ops.utils import Connector, console
+from connector_ops.utils import Connector
 from dagger import Container, DaggerError
-from jinja2 import Environment, PackageLoader, select_autoescape
 from pipelines import sentry_utils
-from pipelines.consts import GCS_PUBLIC_DOMAIN
 from pipelines.helpers.utils import METADATA_FILE_NAME, format_duration, get_exec_result
-from rich.console import Group
-from rich.panel import Panel
 from rich.style import Style
-from rich.table import Table
-from rich.text import Text
-from tabulate import tabulate
 
-from pipelines.models.reports import Report
 
 if TYPE_CHECKING:
     from pipelines.models.contexts import PipelineContext
@@ -44,18 +34,6 @@ class ConnectorWithModifiedFiles(Connector):
     @property
     def has_metadata_change(self) -> bool:
         return any(path.name == METADATA_FILE_NAME for path in self.modified_files)
-
-
-class CIContext(str, Enum):
-    """An enum for Ci context values which can be ["manual", "pull_request", "nightly_builds"]."""
-
-    MANUAL = "manual"
-    PULL_REQUEST = "pull_request"
-    NIGHTLY_BUILDS = "nightly_builds"
-    MASTER = "master"
-
-    def __str__(self) -> str:
-        return self.value
 
 
 class StepStatus(Enum):
@@ -321,152 +299,3 @@ class StepResult:
         return value
 
 
-@dataclass(frozen=True)
-class ConnectorReport(Report):
-    """A dataclass to build connector test reports to share pipelines executions results with the user."""
-
-    @property
-    def report_output_prefix(self) -> str:  # noqa D102
-        return f"{self.pipeline_context.report_output_prefix}/{self.pipeline_context.connector.technical_name}/{self.pipeline_context.connector.version}"
-
-    @property
-    def html_report_file_name(self) -> str:  # noqa D102
-        return self.filename + ".html"
-
-    @property
-    def html_report_remote_storage_key(self) -> str:  # noqa D102
-        return f"{self.report_output_prefix}/{self.html_report_file_name}"
-
-    @property
-    def html_report_url(self) -> str:  # noqa D102
-        return f"{GCS_PUBLIC_DOMAIN}/{self.pipeline_context.ci_report_bucket}/{self.html_report_remote_storage_key}"
-
-    @property
-    def should_be_commented_on_pr(self) -> bool:  # noqa D102
-        return (
-            self.pipeline_context.should_save_report
-            and self.pipeline_context.is_ci
-            and self.pipeline_context.pull_request
-            and self.pipeline_context.PRODUCTION
-        )
-
-    def to_json(self) -> str:
-        """Create a JSON representation of the connector test report.
-
-        Returns:
-            str: The JSON representation of the report.
-        """
-        return json.dumps(
-            {
-                "connector_technical_name": self.pipeline_context.connector.technical_name,
-                "connector_version": self.pipeline_context.connector.version,
-                "run_timestamp": self.created_at.isoformat(),
-                "run_duration": self.run_duration.total_seconds(),
-                "success": self.success,
-                "failed_steps": [s.step.__class__.__name__ for s in self.failed_steps],
-                "successful_steps": [s.step.__class__.__name__ for s in self.successful_steps],
-                "skipped_steps": [s.step.__class__.__name__ for s in self.skipped_steps],
-                "gha_workflow_run_url": self.pipeline_context.gha_workflow_run_url,
-                "pipeline_start_timestamp": self.pipeline_context.pipeline_start_timestamp,
-                "pipeline_end_timestamp": round(self.created_at.timestamp()),
-                "pipeline_duration": round(self.created_at.timestamp()) - self.pipeline_context.pipeline_start_timestamp,
-                "git_branch": self.pipeline_context.git_branch,
-                "git_revision": self.pipeline_context.git_revision,
-                "ci_context": self.pipeline_context.ci_context,
-                "cdk_version": self.pipeline_context.cdk_version,
-                "html_report_url": self.html_report_url,
-                "dagger_cloud_url": self.pipeline_context.dagger_cloud_url,
-            }
-        )
-
-    def post_comment_on_pr(self) -> None:
-        icon_url = f"https://raw.githubusercontent.com/airbytehq/airbyte/{self.pipeline_context.git_revision}/{self.pipeline_context.connector.code_directory}/icon.svg"
-        global_status_emoji = "✅" if self.success else "❌"
-        commit_url = f"{self.pipeline_context.pull_request.html_url}/commits/{self.pipeline_context.git_revision}"
-        markdown_comment = f'## <img src="{icon_url}" width="40" height="40"> {self.pipeline_context.connector.technical_name} test report (commit [`{self.pipeline_context.git_revision[:10]}`]({commit_url})) - {global_status_emoji}\n\n'
-        markdown_comment += f"⏲️  Total pipeline duration: {format_duration(self.run_duration)} \n\n"
-        report_data = [
-            [step_result.step.title, step_result.status.get_emoji()]
-            for step_result in self.steps_results
-            if step_result.status is not StepStatus.SKIPPED
-        ]
-        markdown_comment += tabulate(report_data, headers=["Step", "Result"], tablefmt="pipe") + "\n\n"
-        markdown_comment += f"🔗 [View the logs here]({self.html_report_url})\n\n"
-
-        if self.pipeline_context.dagger_cloud_url:
-            markdown_comment += f"☁️ [View runs for commit in Dagger Cloud]({self.pipeline_context.dagger_cloud_url})\n\n"
-
-        markdown_comment += "*Please note that tests are only run on PR ready for review. Please set your PR to draft mode to not flood the CI engine and upstream service on following commits.*\n"
-        markdown_comment += "**You can run the same pipeline locally on this branch with the [airbyte-ci](https://github.com/airbytehq/airbyte/blob/master/airbyte-ci/connectors/pipelines/README.md) tool with the following command**\n"
-        markdown_comment += f"```bash\nairbyte-ci connectors --name={self.pipeline_context.connector.technical_name} test\n```\n\n"
-        self.pipeline_context.pull_request.create_issue_comment(markdown_comment)
-
-    async def to_html(self) -> str:
-        env = Environment(loader=PackageLoader("pipelines.tests"), autoescape=select_autoescape(), trim_blocks=False, lstrip_blocks=True)
-        template = env.get_template("test_report.html.j2")
-        template.globals["StepStatus"] = StepStatus
-        template.globals["format_duration"] = format_duration
-        local_icon_path = await Path(f"{self.pipeline_context.connector.code_directory}/icon.svg").resolve()
-        template_context = {
-            "connector_name": self.pipeline_context.connector.technical_name,
-            "step_results": self.steps_results,
-            "run_duration": self.run_duration,
-            "created_at": self.created_at.isoformat(),
-            "connector_version": self.pipeline_context.connector.version,
-            "gha_workflow_run_url": None,
-            "dagger_logs_url": None,
-            "git_branch": self.pipeline_context.git_branch,
-            "git_revision": self.pipeline_context.git_revision,
-            "commit_url": None,
-            "icon_url": local_icon_path.as_uri(),
-        }
-
-        if self.pipeline_context.is_ci:
-            template_context["commit_url"] = f"https://github.com/airbytehq/airbyte/commit/{self.pipeline_context.git_revision}"
-            template_context["gha_workflow_run_url"] = self.pipeline_context.gha_workflow_run_url
-            template_context["dagger_logs_url"] = self.pipeline_context.dagger_logs_url
-            template_context["dagger_cloud_url"] = self.pipeline_context.dagger_cloud_url
-            template_context[
-                "icon_url"
-            ] = f"https://raw.githubusercontent.com/airbytehq/airbyte/{self.pipeline_context.git_revision}/{self.pipeline_context.connector.code_directory}/icon.svg"
-        return template.render(template_context)
-
-    async def save(self) -> None:
-        local_html_path = await self.save_local(self.html_report_file_name, await self.to_html())
-        absolute_path = await local_html_path.resolve()
-        if self.pipeline_context.open_report_in_browser:
-            self.pipeline_context.logger.info(f"HTML report saved locally: {absolute_path}")
-            if self.pipeline_context.open_report_in_browser:
-                self.pipeline_context.logger.info("Opening HTML report in browser.")
-                webbrowser.open(absolute_path.as_uri())
-        if self.remote_storage_enabled:
-            await self.save_remote(local_html_path, self.html_report_remote_storage_key, "text/html")
-            self.pipeline_context.logger.info(f"HTML report uploaded to {self.html_report_url}")
-        await super().save()
-
-    def print(self):
-        """Print the test report to the console in a nice way."""
-        connector_name = self.pipeline_context.connector.technical_name
-        main_panel_title = Text(f"{connector_name.upper()} - {self.name}")
-        main_panel_title.stylize(Style(color="blue", bold=True))
-        duration_subtitle = Text(f"⏲️  Total pipeline duration for {connector_name}: {format_duration(self.run_duration)}")
-        step_results_table = Table(title="Steps results")
-        step_results_table.add_column("Step")
-        step_results_table.add_column("Result")
-        step_results_table.add_column("Duration")
-
-        for step_result in self.steps_results:
-            step = Text(step_result.step.title)
-            step.stylize(step_result.status.get_rich_style())
-            result = Text(step_result.status.value)
-            result.stylize(step_result.status.get_rich_style())
-            step_results_table.add_row(step, result, format_duration(step_result.step.run_duration))
-
-        details_instructions = Text("ℹ️  You can find more details with step executions logs in the saved HTML report.")
-        to_render = [step_results_table, details_instructions]
-
-        if self.pipeline_context.dagger_cloud_url:
-            self.pipeline_context.logger.info(f"🔗 View runs for commit in Dagger Cloud: {self.pipeline_context.dagger_cloud_url}")
-
-        main_panel = Panel(Group(*to_render), title=main_panel_title, subtitle=duration_subtitle)
-        console.print(main_panel)
