+++ airbyte-ci/connectors/pipelines/tests/test_steps/test_simple_docker_step.py
@@ -6,8 +6,9 @@ from pathlib import Path
 
 import pytest
 from pipelines.models.contexts import PipelineContext
-from pipelines.models.steps import MountPath, SimpleDockerStep
+from pipelines.models.steps import MountPath
 from pipelines.helpers.utils import get_exec_result
+from pipelines.pipeline.steps.docker import SimpleDockerStep
 
 pytestmark = [
     pytest.mark.anyio,
